requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
psycopg2-binary>=2.9.0
python-multipart>=0.0.6
python-dotenv>=1.0.0
markdown>=3.5.0
PyYaml
sqlalchemy>=2.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-dateutil>=2.8.0